Breaking code up in smaller files to increase CodeCoverage score
